requirements.txt
fastapi
uvicorn
mysql-connector-python
SQLAlchemy
pyjwt
python-jose[cryptography]
python-multipart